requirements.txt
# Telegram Bot
aiogram==3.14.0

# Database
SQLAlchemy==2.0.36

# Scheduler for periodic tasks
APScheduler==3.10.4

# Data processing and visualization
pandas==2.2.3
numpy==2.2.0
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.6.0

# Configuration and environment
python-dotenv==1.0.1

# SSH for backup (if needed)
paramiko==3.5.0

# HTTP requests for Telegram API
requests==2.32.3

# Date/time utilities
python-dateutil==2.9.0.post0
pytz==2024.2
